translator_application_with_flask/flask_requirements.txt
Flask==2.0.2
googletrans==4.0.0-rc1
python-docx==0.8.11
PyPDF2==1.26.0
